Rename modules to differ from functions to appease mypy
